Rename GenericRobotMain
