+++ src/main/java/com/team766/hal/RobotConfiguratorBase.java
@@ -1,5 +1,5 @@
 package com.team766.hal;
 
 public interface RobotConfiguratorBase {
-    GenericRobotMain createRobotMain();
+    GenericRobotMainBase createRobotMain();
 }
