+++ src/main/java/com/team766/hal/simulator/RobotMain.java
@@ -2,7 +2,7 @@ package com.team766.hal.simulator;
 
 import com.team766.config.ConfigFileReader;
 import com.team766.framework.Scheduler;
-import com.team766.hal.GenericRobotMain;
+import com.team766.hal.GenericRobotMainBase;
 import com.team766.hal.RobotProvider;
 import com.team766.hal.RobotSelector;
 import com.team766.logging.LoggerExceptionUtils;
@@ -17,7 +17,7 @@ public class RobotMain {
         VrConnector,
     }
 
-    private GenericRobotMain robot;
+    private GenericRobotMainBase robot;
     private Runnable simulator;
 
     public RobotMain(final Mode mode) {
