+++ src/main/java/com/team766/hal/wpilib/RobotMain.java
@@ -2,7 +2,7 @@ package com.team766.hal.wpilib;
 
 import com.team766.config.ConfigFileReader;
 import com.team766.hal.CanivPoller;
-import com.team766.hal.GenericRobotMain;
+import com.team766.hal.GenericRobotMainBase;
 import com.team766.hal.RobotProvider;
 import com.team766.hal.RobotSelector;
 import com.team766.logging.LoggerExceptionUtils;
@@ -24,7 +24,7 @@ public class RobotMain extends LoggedRobot {
     private static final String USB_CONFIG_FILE = "/U/config/robotConfig.txt";
     private static final String INTERNAL_CONFIG_FILE = "/home/lvuser/robotConfig.txt";
 
-    private GenericRobotMain robot;
+    private GenericRobotMainBase robot;
 
     public static void main(final String... args) {
         Supplier<RobotMain> supplier =
